Add frontend to ssh verification
As stated in https://forum.opennebula.org/t/scp-r-need-password-why/3055/2 in some OpenNebula's operations like clone, scp uses the frontend as the source of the command. In order to avoid key host verification errors, the frontend should be added to the ssh-keyscan command and an ssh connection from the frontend against the frontend itself should be added to the verification steps.