PR
Line 176: changed to 'configureD inter-service connections'. Is it correct to use 'a set set of image parameters'? It makes sense, but perhaps change the second 'set' to another word to avoid confusion (a set list, or something like that?)
Line 199 is not clear to me (perhaps it is for people in the field, though).
Line 334: should this be number 1 instead of 2?