PR
Line 163: 'This address is used by Sunstone interface to redirect to different service FireEdge running within the vOneCloud instance' is unclear. Does it mean that the address is used by the Sunstone interface to redirect FireEdge running within the vOneCloud instance to a different service? This makes sense, but I don't know if that is changing the meaning, so I haven't touched this line. 
Line 264: 'This refinement can be done at import time, defining the size of the network one of the following supported Address Ranges' a word is missing in the second half of this sentence. Do you mean 'defining the size of the network with / according to one of the following...'?