M #-: Review
Line 204: passing by attributes --> I'm not sure if this is the right phrase, but didn't want to change it just in case. 
Line 380: 'It is generally a good idea to place defaults for the KVM-specific attributes' sounds like it needs something else after it. Where you you place them? Or is the default actually the thing that you place (or make)?