PR
Line 178 and line 180 are unclear. Can it be rephrased as something like: 'A hardware configuration as close as possible to that of the Host is preferred' (178)? / 'The configuration with the highest number of threads/cores that fits in the Host is preferred' (180)? I haven't changed anything here as I'm unsure about it.