PR
Lines 13 and 21 are exactly the same. Do you need both of them?
Line 42: 'This tool simplifies the upgrade process of configuration files and always comes in the latest version corresponding to the installed OpenNebula version.' The second part of this sentence is unclear to me. Can it be simplified? Can you say 'comes with the latest version of OpenNebula'?