PR
Line 39: should 'repos' be written in full or is it ok like this?
Line 81: 'any Hosts that use remove monitoring probes' --> this doesn't make sense to me. Should it be remote monitoring probes? 
Line 138: should this be attributed TO the virtual IP address?
Line 208: the second half of this sentence doesn't make much sense.