PR
Line 50: I've changed this slightly to add an object (you), because we need one with 'allow'. Is this ok?
Line 82: should this be schedule actions or scheduled actions? Scheduled actions works if they are already scheduled, but schedule actions sounds like a feature you can implement.
Line 120: provision operations has/have been implemented? Have if these are separate operations or has if it is a feature/tool called provision operations. I've not changed it here, but wanted to point it out.
Line 121: 'multiple same Hosts' --> 'multiple identical Hosts'. Is this change ok? It sounds better than 'same'.
Line 123: has/have been added. Either could work (the difference is the same as in line 120).