PR
Line 154 and line 204: 'we restrict access only from IP address...' means that you are only blocking access from that IP address. Do you mean that you an only access the VNC from that IP address, and no other? If so, I'd change it to: 'In this example, we restrict access from all IP addresses except 192.168.0.1'.
Line 350 in table: the end of the sentence is missing and needs to be added.